Update pull aws gfs script to get 0-hour data from NOMADS.

+++ components/scripts/common/pull_aws_s3_gfs.ksh
@@ -4,6 +4,18 @@
 # README File:  https://docs.opendata.aws/noaa-gfs-pds/readme.html
 # GFS AWS Data: http://awsopendata.s3-website-us-west-2.amazonaws.com/noaa-gfs
 
+# This script attempts to wget all URL's from pull_aws_s3_gfs_tmpl.txt.
+# However, not all files actually exist for each forecast hour.
+# The list of URL's was generated with these commands:
+#
+#   rm pull_aws_s3_gfs_tmpl.txt
+#   for VAR in `aws s3 ls s3://noaa-gfs-pds | sed -r 's/^ +PRE //g' | cut -d '/' -f1`; do
+#     for LEV in `aws s3 ls s3://noaa-gfs-pds/$VAR/ | sed -r 's/^ +PRE //g' | sed -r 's/ /%20/g' | cut -d '/' -f1`; do
+#       echo http://noaa-gfs-pds.s3.amazonaws.com/${VAR}/${LEV}/{INIT_YMD}/{INIT_H}00/{FHR}
+#       echo http://noaa-gfs-pds.s3.amazonaws.com/${VAR}/${LEV}/{INIT_YMD}/{INIT_H}00/{FHR} >> pull_aws_s3_gfs_tmpl.txt
+#     done
+#  done
+
 typeset -Z3 CUR_FHR
 
 # Check arguments
@@ -15,6 +27,8 @@ fi
 INIT_YMDH=$1
 MAX_FHR=$2
 FHR_INC=3
+
+WGRIB2=/home/ec2-user/utilities/grib2/wgrib2/wgrib2
 
 GFS_URL_TMPL=`dirname $0`/pull_aws_s3_gfs_tmpl.txt
 
@@ -29,26 +43,43 @@ while [[ $CUR_FHR -le $MAX_FHR ]]; do
   echo "PULLING: GFS data for the $INIT_YMDH initialization $CUR_FHR forecast hour."
   CUR_TMP_DIR=gfs_f${CUR_FHR}
   CUR_FILE_LIST=aws_s3_gfs_files_${INIT_YMDH}_f${CUR_FHR}.txt
+  CUR_GRB_FILE=gfs_0p25_${INIT_YMD}_${INIT_H}00_${CUR_FHR}.grb2
 
-  # Delete directory in case it already exists
-  rm -rf $CUR_TMP_DIR
-  mkdir $CUR_TMP_DIR
+  # Pull 0-hour forecast from NOMADS since s3 does not contain them
+  if [[ $CUR_FHR -eq 0 ]]; then 
 
-  # Customize the file list for the current times
-  cat $GFS_URL_TMPL | sed "s/{INIT_YMD}/${INIT_YMD}/g" | \
-     sed "s/{INIT_H}/${INIT_H}/g" | sed "s/{FHR}/${CUR_FHR}/g" > \
-     $CUR_TMP_DIR/$CUR_FILE_LIST
-  cd $CUR_TMP_DIR
+    echo "RUNNING: time wget -q https://nomads.ncep.noaa.gov/pub/data/nccf/com/gfs/prod/gfs.${INIT_YMD}/${INIT_H}/gfs.t${INIT_H}z.pgrb2.0p25.f${CUR_FHR}"
+    time wget -q https://nomads.ncep.noaa.gov/pub/data/nccf/com/gfs/prod/gfs.${INIT_YMD}/${INIT_H}/gfs.t${INIT_H}z.pgrb2.0p25.f${CUR_FHR}
 
-  # Pull the files
-  wget -i $CUR_FILE_LIST
-  cd ..
+    # Rename this file
+    mv gfs.t${INIT_H}z.pgrb2.0p25.f${CUR_FHR} $CUR_GRB_FILE
 
-  # Concatenate all the files for this time
-  cat $CUR_TMP_DIR/${CUR_FHR}* > gfs_4_${INIT_YMD}_${INIT_H}00_${CUR_FHR}.grb2
+  else
 
-  # Delete the temporary directory 
-  rm -rf $CUR_TMP_DIR
+    # Delete directory in case it already exists
+    rm -rf $CUR_TMP_DIR
+    mkdir $CUR_TMP_DIR
+
+    # Customize the file list for the current times
+    cat $GFS_URL_TMPL | sed "s/{INIT_YMD}/${INIT_YMD}/g" | \
+       sed "s/{INIT_H}/${INIT_H}/g" | sed "s/{FHR}/${CUR_FHR}/g" > \
+       $CUR_TMP_DIR/$CUR_FILE_LIST
+    cd $CUR_TMP_DIR
+
+    # Pull the files
+    echo "RUNNING: time wget -q -i $CUR_FILE_LIST"
+    time wget -q -i $CUR_FILE_LIST
+    cd ..
+
+    # Concatenate all the files for this time
+    cat $CUR_TMP_DIR/${CUR_FHR}* > $CUR_GRB_FILE
+
+    # Delete the temporary directory 
+    rm -rf $CUR_TMP_DIR
+
+  fi
+
+  echo "WRITING: $CUR_GRB_FILE with `$WGRIB2 $CUR_GRB_FILE | wc -l` records."
 
   # Increment the forecast hour
   CUR_FHR=$(($CUR_FHR + $FHR_INC))
